requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.0.9
django-allauth==0.42.0
django-anymail==7.2.1
django-djconfig==0.10.0
django-haystack==3.0b1
django-infinite-scroll-pagination==1.1.0
django-spirit==0.8.0
idna==2.10
mistune==0.8.4
oauthlib==3.1.0
olefile==0.46
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
Whoosh==2.7.4
